backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
httpx==0.25.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.6.1
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
lxml==4.9.3
python-dateutil==2.8.2
